backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
pyotp==2.9.0
psycopg[binary]==3.1.18
django-cors-headers==4.3.1
whitenoise==6.6.0
gunicorn==22.0.0
